Remove THIRD_PARTY_APPS and define configuration per third-party app
Its practical and tidy to group configurations per app.

A lot of third party apps require a significant amount of configuration,
it's convenient to be able to look at the configuration
of an app in previous project you used the app in. The existance of
THIRD_PARTY_APPS then violates DRY.

There are THIRD_PARTY apps like South where an entry in INSTALLED_APPS
is the only configuration. I believe it is more important to group
the configurations per app together at the cost of having a few extra
commented lines than not grouping third party configurations at all.